transferring NewRecordDetailViewController code over to RecordDetailViewController, connecting segue in storyboard
